+++ internal/delivery/http/router/auth_route.go
@@ -2,13 +2,11 @@ package router
 
 import (
 	"github.com/afifudin23/absensi-king-royal-api/internal/delivery/http/handler"
-	"github.com/afifudin23/absensi-king-royal-api/internal/service"
 	"github.com/gin-gonic/gin"
 )
 
 func registerAuthRoutes(rg *gin.RouterGroup) {
-	authService := service.NewAuthService()
-	authHandler := handler.NewAuthHandler(authService)
+	authHandler := handler.NewAuthHandler()
 
 	auth := rg.Group("/auth")
 	auth.POST("/register", authHandler.Register)