+++ internal/delivery/http/router/router.go
@@ -1,11 +1,18 @@
 package router
 
 import (
+	"github.com/afifudin23/absensi-king-royal-api/internal/middleware"
 	"github.com/gin-gonic/gin"
 )
 
 func New() *gin.Engine {
-	r := gin.Default()
+	r := gin.New()
+
+	r.Use(
+		gin.Logger(),
+		middleware.RecoveryMiddleware(),
+		middleware.ErrorMiddleware(),
+	)
 
 	api := r.Group("/api")
 	v1 := api.Group("/v1")
@@ -13,6 +20,7 @@ func New() *gin.Engine {
 	registerHealthRoutes(v1)
 	registerRootRoutes(v1)
 	registerAuthRoutes(v1)
+	registerUserRouter(v1)
 
 	return r
 }